- Cambiando el manejo de las tarjetas en Inicio y Resultados - Agregando un formulario para reservas - Agregando una pantalla de carga genérica e implmentandola en una prueba en el fxml de Resultados

+++ src/main/java/com/example/hotel/dominio/Hotel.java
@@ -43,7 +43,7 @@ public class Hotel {
   private String imagenUrl;
   
   @Column(nullable =  false)
-  private short estrellas;
+  private byte estrellas;
 
   @Column(length = 150, nullable =  false)
   private String nombre;
